AnalyszeDataWithAI/requirements.txt
e2b-code-interpreter
openai
python-dotenv
